feat: add SEO support with robots and sitemap generation
- Implemented robots.ts to define crawling rules and sitemap URL.
- Created sitemap.ts to list key public pages for search engines.
- Updated scoring methodology page to use new SEO components and structured data.
- Added JsonLd component for structured data representation.
- Refactored scoring methodology page to separate client-side logic.
- Enhanced app header with a link to the home page.
- Improved compare form with additional header information.
- Created home-page-client component for better state management and user experience.
- Added tests for SEO helpers and routes to ensure proper functionality.
- Updated localization files for new strings related to the changes.
- Added Open Graph image for better social media sharing.
