+++ app/layout.tsx
@@ -1,5 +1,6 @@
 import "./globals.css";
 import type { ReactNode } from "react";
+import type { Metadata } from "next";
 import { cookies, headers } from "next/headers";
 import Script from "next/script";
 import {
@@ -10,6 +11,7 @@ import {
   parseAcceptLanguage,
   supportedLocales,
 } from "@/lib/i18n-core";
+import { getMetadataBase, toAbsoluteUrl } from "@/lib/seo";
 import Providers from "./providers";
 
 const themeInitScript = `
@@ -30,9 +32,64 @@ const themeInitScript = `
   } catch {}
 `;
 
-export const metadata = {
-  title: "DevImpact",
-  description: "GitHub user scoring",
+export const metadata: Metadata = {
+  title: {
+    default: "DevImpact | GitHub Developer Comparison & Open-Source Impact Scoring",
+    template: "%s | DevImpact",
+  },
+  description:
+    "Compare GitHub developers by repository impact, merged external pull requests, and community contribution signals with transparent scoring.",
+  keywords: [
+    "github developer comparison",
+    "open source impact score",
+    "github repository analytics",
+    "pull request impact",
+    "developer ranking tool",
+    "devimpact",
+  ],
+  authors: [{ name: "DevImpact Team" }],
+  creator: "DevImpact",
+  publisher: "DevImpact",
+  metadataBase: getMetadataBase(),
+  alternates: {
+    canonical: "/",
+  },
+  openGraph: {
+    type: "website",
+    title: "DevImpact | GitHub Developer Comparison & Open-Source Impact Scoring",
+    description:
+      "Compare GitHub developers with transparent repo, PR, and community contribution scoring.",
+    url: "/",
+    siteName: "DevImpact",
+    images: [
+      {
+        url: toAbsoluteUrl("/og-image.svg"),
+        width: 1200,
+        height: 630,
+        alt: "DevImpact GitHub developer comparison dashboard preview",
+      },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "DevImpact | GitHub Developer Comparison",
+    description:
+      "Compare open-source impact using repository, pull request, and community contribution signals.",
+    images: [toAbsoluteUrl("/og-image.svg")],
+  },
+  robots: {
+    index: true,
+    follow: true,
+    nocache: false,
+    googleBot: {
+      index: true,
+      follow: true,
+      "max-image-preview": "large",
+      "max-snippet": -1,
+      "max-video-preview": -1,
+    },
+  },
+  manifest: "/manifest.webmanifest",
   icons: {
     icon: "/logo.svg",
     shortcut: "/logo.svg",
