+++ app/manifest.ts
@@ -1,0 +1,22 @@
+import type { MetadataRoute } from "next";
+import { SITE_NAME, SITE_SHORT_NAME } from "@/lib/seo";
+
+export default function manifest(): MetadataRoute.Manifest {
+  return {
+    name: SITE_NAME,
+    short_name: SITE_SHORT_NAME,
+    description:
+      "Compare GitHub developers with transparent open-source impact scoring.",
+    start_url: "/",
+    display: "standalone",
+    background_color: "#ffffff",
+    theme_color: "#0ea5e9",
+    icons: [
+      {
+        src: "/logo.svg",
+        sizes: "any",
+        type: "image/svg+xml",
+      },
+    ],
+  };
+}
