+++ app/robots.ts
@@ -1,0 +1,18 @@
+import type { MetadataRoute } from "next";
+import { getSiteUrl } from "@/lib/seo";
+
+export default function robots(): MetadataRoute.Robots {
+  const siteUrl = getSiteUrl();
+
+  return {
+    rules: [
+      {
+        userAgent: "*",
+        allow: ["/", "/scoring-methodology"],
+        disallow: ["/api/", "/_next/", "/private/"],
+      },
+    ],
+    sitemap: `${siteUrl}/sitemap.xml`,
+    host: siteUrl,
+  };
+}
