+++ app/sitemap.ts
@@ -1,0 +1,22 @@
+import type { MetadataRoute } from "next";
+import { getSiteUrl } from "@/lib/seo";
+
+export default function sitemap(): MetadataRoute.Sitemap {
+  const baseUrl = getSiteUrl();
+  const now = new Date();
+
+  return [
+    {
+      url: `${baseUrl}/`,
+      lastModified: now,
+      changeFrequency: "weekly",
+      priority: 1,
+    },
+    {
+      url: `${baseUrl}/scoring-methodology`,
+      lastModified: now,
+      changeFrequency: "monthly",
+      priority: 0.7,
+    },
+  ];
+}
