+++ components/app-header.tsx
@@ -1,3 +1,4 @@
+import Link from "next/link";
 import { BrandLogo } from "@/components/brand-logo";
 import { LanguageSwitcher } from "@/components/language-switcher";
 import { ThemeToggle } from "@/components/theme-toggle";
@@ -6,7 +7,9 @@ export function AppHeader() {
   return (
     <header className="sticky top-0 z-50 w-full border-b border-border bg-background/95 backdrop-blur supports-[backdrop-filter]:bg-background/60">
       <div className="container m-auto flex h-20 max-w-7xl items-center justify-between px-4 sm:h-24">
-        <BrandLogo priority size="md" className="shrink-0" />
+        <Link href="/" className="shrink-0" aria-label="DevImpact home">
+          <BrandLogo priority size="md" />
+        </Link>
 
         <div className="flex gap-4">
           <LanguageSwitcher />
