+++ locales/ar.json
@@ -79,6 +79,7 @@
   "footer.tagline": "رؤية أوضح. تقييم أسرع.",
   "form.compare": "قارن",
   "form.compare.ing": "جاري المقارنة...",
+  "form.header.eyebrow": "أثر المطور",
   "form.enterTwo": "أدخل اسمي مستخدمين لمقارنة المقاييس",
   "form.insights.toggle": "توليد رؤى",
   "form.languages.clear": "مسح اللغات",
